refactor(component): move `window` event listener
