+++ src/DrumPad.tsx
@@ -1,16 +1,28 @@
 import type { ReactElement } from "react";
 
-import type { DrumPadLabel } from "./types";
+import type { DrumPadLabel, DrumPadHotkey } from "./types";
 
-import { DRUM_PADS } from "./constants";
+import { DRUM_PADS, DRUM_PAD_HOTKEYS } from "./constants";
 
 import to_kebab_case from "./helpers/to_kebab_case";
+import to_title_case from "./helpers/to_title_case";
+
+window.addEventListener("keydown", (event: KeyboardEvent) => {
+	if (DRUM_PAD_HOTKEYS.indexOf(event.key.toLocaleLowerCase() as DrumPadHotkey) === -1) return;
+
+	const audio = document.querySelector(`audio#${event.key.toUpperCase()}`) as HTMLAudioElement;
+	const drum_pad = audio.parentElement as HTMLButtonElement;
+	const display = document.querySelector("#display") as HTMLHeadingElement;
+
+	audio.play().then((): void => {
+		display.innerText = to_title_case(drum_pad.id).toLocaleUpperCase();
+	}).catch(error => console.error(error));
+});
 
 interface Props {
 	label: DrumPadLabel
 	set_audio_to_play: React.Dispatch<React.SetStateAction<DrumPadLabel | "">>
 }
-
 
 export default function DrumPad({ label, set_audio_to_play }: Props): ReactElement {
 	function handle_click(event: React.MouseEvent<HTMLButtonElement>, label: DrumPadLabel): void {
